+++ example/simple_cache_example.dart
@@ -1,7 +1,7 @@
 import 'package:dcache/dcache.dart';
 
 void main() {
-  Cache c = new SimpleCache<String, int>(storage: new SimpleStorage<String, int>(size: 20));
+  Cache c = new SimpleCache<String, int>(storage: new InMemoryStorage<String, int>(20));
 
   c.set("key", 42);
   print(c.get("key")); // 42