+++ test/simple_test.dart
@@ -4,30 +4,30 @@ import 'dart:io';
 import 'package:dcache/dcache.dart';
 
 void main() {
-  test("Test cache initialization", () {
-    Cache cache = new SimpleCache<int, int>(storage: new SimpleStorage(size: 20));
+  test('Test cache initialization', () {
+    Cache cache = SimpleCache<int, int>(storage: InMemoryStorage(20));
     expect(cache, isNotNull);
   });
   //
-  test("Test simple insert/get", () {
-    Cache c = new SimpleCache<String, int>(storage: new SimpleStorage(size: 20));
+  test('Test simple insert/get', () {
+    Cache c = SimpleCache<String, int>(storage: InMemoryStorage(20));
 
-    c.set("key", 42);
-    expect(c.get("key"), equals(42));
+    c.set('key', 42);
+    expect(c.get('key'), equals(42));
   });
-  test("Test simple loader function", () {
+  test('Test simple loader function', () {
     Cache<int, int> c =
-        new SimpleCache<int, int>(storage: new SimpleStorage(size: 20))
+        SimpleCache<int, int>(storage: InMemoryStorage(20))
           ..loader = (int k, _) => k * 10;
     expect(c.get(4), equals(40));
     expect(c.get(5), equals(50));
   });
-  test("Test simple loader function", () {
+  test('Test simple loader function', () {
     Cache<int, int> c =
-        new SimpleCache<int, int>(storage: new SimpleStorage(size: 20))
+        SimpleCache<int, int>(storage: InMemoryStorage(20))
           ..syncLoading = true
           ..expiration = const Duration(milliseconds: 200)
-          ..loader = (int k, int oldValue) {
+          ..loader = (int k, int? oldValue) {
             oldValue ??= k;
             var v = oldValue * 10;
             return v;
@@ -39,12 +39,12 @@ void main() {
 
     expect(c.get(4), equals(400));
   });
-  test("Test simple loader function", () {
+  test('Test simple loader function', () {
     Cache<int, int> c =
-        new SimpleCache<int, int>(storage: new SimpleStorage(size: 20))
+        SimpleCache<int, int>(storage: InMemoryStorage(20))
           ..syncLoading = false
           ..expiration = const Duration(milliseconds: 200)
-          ..loader = (int k, int oldValue) {
+          ..loader = (int k, int? oldValue) {
             oldValue ??= k;
             var v = oldValue * 10;
             return v;
@@ -56,9 +56,25 @@ void main() {
 
     expect(c.get(4), equals(40));
   });
-  test("Test simple eviction", () {
+  test('Test simple async loader function', () async {
     Cache<int, int> c =
-        new SimpleCache<int, int>(storage: new SimpleStorage(size: 3))
+    SimpleCache<int, int>(storage: InMemoryStorage(20))
+      ..syncLoading = false
+      ..expiration = const Duration(milliseconds: 200)
+      ..loader = (int k, int? oldValue) async {
+        oldValue ??= k;
+        var v = oldValue * 10;
+        return v;
+      };
+
+
+    expect(c.get(4), equals(null));
+    await Future<dynamic>.delayed(Duration(seconds: 1));
+    expect(c.get(4), equals(40));
+  });
+  test('Test simple eviction', () {
+    Cache<int, int> c =
+        SimpleCache<int, int>(storage: InMemoryStorage(3))
           ..loader = (int k, _) {
             return k * 10;
           };
@@ -69,9 +85,9 @@ void main() {
     expect(c.get(7), equals(70));
     expect(c.containsKey(4), equals(false));
   });
-  test("Test LRU eviction", () {
+  test('Test LRU eviction', () async {
     Cache<int, int> c =
-        new LruCache<int, int>(storage: new SimpleStorage(size: 3))
+        LruCache<int, int>(storage: InMemoryStorage(3))
           ..loader = (int k, _) {
             return k * 10;
           };
@@ -79,14 +95,15 @@ void main() {
     expect(c.get(4), equals(40));
     expect(c.get(5), equals(50));
     expect(c.get(6), equals(60));
+    await Future<dynamic>.delayed(Duration(seconds: 1));
     expect(c.get(4), equals(40));
     expect(c.get(6), equals(60));
     expect(c.get(7), equals(70));
     expect(c.containsKey(5), equals(false));
   });
-  test("Test LFU eviction", () {
+  test('Test LFU eviction', () {
     Cache<int, int> c =
-        new LfuCache<int, int>(storage: new SimpleStorage(size: 3))
+        LfuCache<int, int>(storage: InMemoryStorage(3))
           ..loader = (int k, _) {
             return k * 10;
           };