+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+## 0.4.0
+
+* Allow async loader function [issue-11](https://github.com/platelk/dcache/issues/11)
+* Change logic to handle eviction on common cache logic [issue-10](https://github.com/platelk/dcache/issues/10)
+* Null Safety compliant for dart 2.12
+* Update dependencies
+
 ## 0.3.0
 
 * Add `OnEvit` option to catch eviction and do logic on it (thx to  [mikes222](https://github.com/mikes222))